requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-multipart==0.0.5
bcrypt==3.2.0
passlib==1.7.4
python-jose==3.3.0
cryptography==36.0.0
